+++ src/main/java/roomescape/login/controller/LoginController.java
@@ -1,4 +1,4 @@
-package roomescape.login;
+package roomescape.login.controller;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
@@ -9,6 +9,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import roomescape.login.dto.LoginRequest;
+import roomescape.login.dto.LoginResponse;
+import roomescape.login.service.LoginService;
 
 @Controller
 @RequestMapping("/login")
